:bear: turn components into ES6 classes
... with their own JS files

:bear: turn NotFound and StorePicker into ES6 classes

:bear: turn Header into an ES6 class

:bear: turn Fish into an ES6 class

:bear: turn AddFishForm into an ES6 class

:bear: turn Order into an ES6 class

:bear: turn Inventory into an ES6 class

:bear: missing semi-colon
