+++ 21 - ES6 Modules/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "catch-of-the-day",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "autobind-decorator": "1.3.2",
+    "babel": "5.8.23",
+    "babelify": "6.4.0",
+    "browser-sync": "2.9.11",
+    "browserify": "11.2.0",
+    "connect-history-api-fallback": "1.1.0",
+    "firebase": "2.3.1",
+    "gulp": "3.9.0",
+    "gulp-autoprefixer": "3.1.0",
+    "gulp-notify": "2.2.0",
+    "gulp-rename": "1.2.2",
+    "gulp-stylus": "2.1.0",
+    "gulp-uglify": "1.4.2",
+    "gulp-util": "3.0.7",
+    "history": "1.12.5",
+    "install": "0.1.8",
+    "re-base": "1.2.1",
+    "react": "0.14.0",
+    "react-addons-css-transition-group": "0.14.0",
+    "react-catalyst": "0.3.0",
+    "react-dom": "0.14.0",
+    "react-mixin": "3.0.0",
+    "react-router": "^1.0.0-rc3",
+    "vinyl-buffer": "1.0.0",
+    "vinyl-source-stream": "1.1.0",
+    "watchify": "3.4.0"
+  }
+}
