+++ 21 - ES6 Modules/scripts/components/NotFound.js
@@ -1,0 +1,13 @@
+/*
+ Not Found
+ */
+
+import React from 'react';
+
+var NotFound = React.createClass({
+  render : function() {
+    return <h1>Not Found!</h1>
+  }
+});
+
+export default NotFound;
